wip: tests: add CRL mTLS test.
This commit adds a simple test CRL (`testdata/test.crl.pem`) that lists
the `testdata/localhost/cert.pem` certificate as revoked, but _not_ the
`testdata/example.com/cert.pem` certificate.

The `client-server.py` integration test driver is then updated with
a suite that will start the server binary in a mode that requires mTLS,
and that loads the test crl. Two connection attempts are made with the
client binary: one using the `example.com` client cert that isn't
expected to error, and one using the `localhost` client cert that _is_
expected to error (since it's revoked).