server: make server config builder set persistence void
The `rustls_server_config_builder_set_persistence` fn previously
returned a `rustls_result`, but the only possible error was
`rustls_result::NullParameter` if any of the three arguments were
`NULL`. In this circumstance we prefer to return early, dong nothing
with side effects. This commit changes
`rustls_server_config_builder_set_persistence` to be `void` and updates
the documentation accordingly.