ci: restore full integration coverage for Windows
Previously we ran the static libs integration test in
`Makefile.windows`, and not the CMake build, because they disagreed on
output. We since removed the `Makefile.windows` build, but forgot to
update one of the CMake builds to do that job. This commit fixes this by
having the release CMake Windows build run all the ignored tests.