connection: avoid `rustls_supported_ciphersuite` ptr
Previously the `rustls_connection_get_negotiated_ciphersuite` function
returned a pointer to a `rustls_supported_ciphersuite`. This commit
changes that function to only return the identifier int. A new
`rustls_connection_get_negotiated_ciphersuite_name` function is added
for getting the negotiated ciphersuite name as a `rustls_str`.

We want to avoid returning a `rustls_supported_ciphersuite` here because
this type is both the _implementation_ of a ciphersuite, and metadata
such as the name/ID. Getting a handle to the implementation for a given
connection requires iterating the `ALL_CIPHER_SUITES` array to find the
matching ciphersuite impl, and this is only workable when the available
ciphersuites is a fixed quantity. Soon we will support customizing the
cryptography provider, complicating this design greatly.

Functionally, the only thing a caller wants to do in this circumstance
is find the negotiated ciphersuite ID or name. To avoid the complication
discussed above we can simply return this information directly and avoid
the need to find the full fledged ciphersuite implementation at
connection time.