crypto_provider: add signing key loading support
This commit adds a new type, `rustls_signing_key`, that represents
a `&dyn SigningKey` loaded by a `rustls_crypto_provider`.

A new `rustls_crypto_provider_load_key` fn is added to create
a `rustls_signing_key` from a pointer to a `rustls_crypto_provider`, and
PEM content in-memory.

Wiring this up will be done in a subsequent commit.