cipher: use provider to load signing keys
This breaks an unconditional dependency on `*ring*` for loading
certified key private keys.

The `common.c` function for loading a certified key is updated to
accept a crypto provider pointer to use for loading a signing key.

The `server.c` and `client.c` test binaries are then updated to provide
the default provider to the common function as appropriate.