cipher: convert server/client verifiers to provider
This breaks an unconditional dep on `*ring*` for both verifiers.

The client/server test binaries do not require any update in this case
since they are using the APIs that assume a process-wide default crypto
provider has been set and we've already done that w/ ring early in
`main` in both cases.