tests: support customizing supported ciphersuite
This commit updates both `client.c` and `server.c` to respect a new
`RUSTLS_CIPHERSUITE` env var. When set, the process-default cryptography
provider's supported ciphersuites will be reduced to _just_ the one
specified by name in the env var.

The `client_server.rs` integration test is then updated to start
a server that only supports one ciphersuite. Two clients are created,
one with a matching ciphersuite and one without. We use each client to
connect to the server and assert only the expected one with matching
ciphersuite support works.