client: remove NoneVerifier
Now that the `rustls_client_config_builder_build()` fn is fallible it
makes more sense to return an error when the required server certificate
verifier hasn't been set instead of using `NoneVerifier` and failing all
certificate validations.

This commit removes the `NoneVerifier` and updates the tests that were
building a client config without specifying a verifier to use the
platform verifier instead. A new unit test is added that ensures the
correct error is returned when a config is built without a verifier.