client: make config builder building fallible
Previously the `rustls_client_config_builder_build` function was
infallible and returned a `rustls_client_config` instance without any
chance to communicate errors. This commit updates this function to
instead return `rustls_result` and use an out parameter for the
`rustls_client_config`.

Having this function unable to return a detailed error has a number of
knock-on effects we want to address:

* If no server certificate verifier has been configured the previous
  implementation isn't able to communicate that and instead configures
  a `NoneVerifier` - this ends up pushing failures to the time of
  certificate verification, making for a subpar user experience. If the
  user intended to disable certificate validation they need to do so
  with a verifier that won't error on use. If the user intended to
  configure certificate validation but something went wrong, we've made
  debugging harder.
* Shortly we will allow customizing the crypto provider used for the
  configuration and may need to error if no suitable provider has been
  configured.

The `client.c` example and `client.rs` unit tests are both updated to
use the new out-parameter based API.