connection: add API for queued traffic key refresh
Adds a `rustls_connection_refresh_traffic_keys()` fn for queuing
a traffic key refresh, e.g. because you know the connection is about to
be idle for a long time and you wish to roll keys ahead of this.

There's not a great place to use this from `client.c` or `server.c` so
for now I've added this API without integration test coverage.