add opt-in FIPS feature, CI coverage
Using `make FIPS=true` with the Makefiles, or `cmake -DFIPS="true" -S
. -B build` with the Windows cmake build will activate the `aws-lc-rs`
feature of `rustls-ffi`, and the `rustls/fips` feature of Rustls.

On MacOS and Windows this requires some additional build tooling (Golang
and Ninja). See the rustls manual[0] and the aws-lc-rs-fips-sys crate[1]
for more information.

Note presently the Mac and Windows FIPS-enabled builds fail with
unresolved symbol errors when building the client/server examples. A fix
is TBD.

[0]: https://docs.rs/rustls/latest/rustls/manual/_06_fips/index.html
[1]: https://crates.io/crates/aws-lc-fips-sys