tests/client: support ECH GREASE
As of this commit if the `RUSTLS_ECH_GREASE` env var is set then
the client will use the first HPKE suite supported by aws-lc-rs to
generate a random HPKE public key. The `client_config_builder` is then
customized to enable ECH in GREASE mode using the suite and generated
pubkey.

If the aws-lc-rs crypto provider is not available and
`RUSTLS_ECH_GREASE` is set, then an error is returned.