cmake: set default CMAKE_BUILD_TYPE when sensible
This is largely in preparation for supporting MacOS and Linux CMake
builds where the default for the C/C++ tooling is annoying to deal with.
It is a no-op for Windows/IDE based CMake builds that use a "Multi
config" setup.

See the diff's comment for more detail on why this is trickier than it
has any right to be.