tests: fix cast for ALPN data print
The `rustls_connection_get_alpn_protocol()` function takes an out param
of type `const uint8_t*` for the ALPN data and an out param of type
`size_t` for the length. We were using these with `printf()` for
a `%.*s` specifier. This requires that we cast the len to `int` (already
done), as well as casting the `const uint8_t*` to `const char*` (done as
of this commit).

Along the way, move the declarations of the two out vars closer to the
only code that uses them.