tools/docgen: simplify comment_and_requirement
Simplify the logic by handling the case where `prev` is a comment that
may/may not have a feature req previous to the comment up-front.

Then, when we're looking at a `prev` that wasn't a comment, but was
`kind == expression_statement` we can recurse with `prev_prev` to handle
it as a comment that may/may not have a feature req previous to itself.

Lastly, if we know we don't have a comment and we don't have an
expression statement, try to see if we have a bare feature requirement.
We know cbindgen won't make a comment ahead of these and so can
disregard `prev_prev` for this case.